refine trie search benchmark
