+++ tools/app/trie_search_cost.chart
@@ -1,0 +1,12 @@
+cost of trie search with existing & existent key:
+┌────────┬────────┬────────────────────┬───────────────────────┐
+│ KeyCnt │ KeyLen │ ExsitingKeyNsPerOp │ NonexsitentKeyNsPerOp │
+├────────┼────────┼────────────────────┼───────────────────────┤
+│ 1      │ 1024   │ 116                │ 85                    │
+│ 10     │ 1024   │ 119                │ 96                    │
+│ 100    │ 1024   │ 224                │ 197                   │
+│ 1000   │ 1024   │ 222                │ 206                   │
+│ 1000   │ 512    │ 272                │ 241                   │
+│ 1000   │ 256    │ 222                │ 190                   │
+└────────┴────────┴────────────────────┴───────────────────────┘
+
